Fix critical issues from code reviews: race condition, pagination, precision
CRITICAL FIXES:
1. Race Condition in Status Updates (High Priority)
   - Fix lock release in no-op case: commit transaction before returning
   - Prevents premature lock release that could allow concurrent updates
   - Ensures transaction completes before lock is released

2. Incorrect Pagination Totals (Medium Priority)
   - Add count_ml_analyses_for_image() and count_ml_annotations()
   - Return actual total record counts instead of len(results)
   - Fixes pagination UI showing incorrect totals when skip/limit applied

3. Confidence Numeric Precision Bug (Medium Priority)
   - Change confidence from Numeric(5,4) to Float
   - Numeric(5,4) only allows max 9.9999, not 1.0
   - Add Alembic migration for schema change
   - Prevents precision errors for confidence scores

These fixes address the top issues identified in PR code reviews.

+++ backend/core/models.py
@@ -199,7 +199,7 @@ class MLAnnotation(Base):
     analysis_id = Column(UUID(as_uuid=True), ForeignKey("ml_analyses.id", ondelete="CASCADE"), nullable=False, index=True)
     annotation_type = Column(String(50), nullable=False, index=True)  # classification, bounding_box, heatmap, segmentation
     class_name = Column(String(255), nullable=True)
-    confidence = Column(Numeric(5, 4), nullable=True)
+    confidence = Column(Float, nullable=True)  # Confidence score 0.0-1.0
     data = Column(JSON, nullable=False)  # dynamic payload: coordinates, arrays, etc.
     storage_path = Column(String(1024), nullable=True)  # pointer to artifact in object storage
     ordering = Column(Integer, nullable=True)
